Update Debian packaging to depend on site-builder package.
This change causes the site build to use the site-builder
package installed as a Debian package. The site-builder
package is no longer included in the node_modules directory
include in the gov-site package.

It also excludes devDependencies from the generated Debian
package. The devDependencies are not loaded when running
build, publish or checklinks.