Configure cargo plugin to work mygov-parent pom
Hippo documentation assumes that users will use the hippo-cms7-release
project as a parent pom. This allows projects that use Hippo to inherit
plugin configuration using pluginManagement. However, this would prevent
the use of the mygov-parent pom.

To use the mygov-parent pom, plugins usually configured in the
hippo-cms7-release and hippo-cms7-project poms must be configured.