MGS-4160 Configure vhosts for use with load balancers
When deploying on a full size environment, the requests for pages
are proxies through the internal load balancer (ALB). The ALB sets
X-Forwarded-Proto and X-Forwarded-Port HTTP headers based on the
request from the web server to the load balancer, rather than from
browser to external load balancer.

Hippo uses these headers to determine whether to return content
or a redirect to the canonical protocol (https) and port (443).

Therefore, to prevent Hippo from always returning redirects,
this change sets the `schemeagnostic` property on the vhost
to true, and `showport` to true. The former allows traffic via
the load balancer to be HTTP and the latter allows it to use
any port.