MGS-4426 Improve indexing configuration override
Overriding the built-in indexing behaviour by including a file in our
jar may prove fragile (e.g. changes in pom files altering the classpath,
or if the directory structure changes in the future). Instead, this file
can be copied to the installation directory and referenced explicitly.

However, since this reference is handled by the repository.xml provided
by our puppetry, this file should still be present and included in the
jar so that a local cargo run can use it to override the default
behaviour.