MGS-4764 Fix bug that caused /index to be redirected to the web archive
We use a tree structure of all known historical news items to decide if a url should be redirected to the
web archive.  However this had a bug such that any url that started with an archived slug would be redirected.
This includes /index which caused the wrong canonical url of the landing page to go to the archive.