MGS-4806 MGS-4812 Improve publication link processor
Switched from a SQL query with multiple wildcards to an XPath query. The
former was inefficient and resulted in high CPU and memory usage when
under load.

Use ISO9075 encoding method to escape paths in XPath query. This avoids
log errors for JCR nodes that start with a digit (since this isn't valid
XML), and sanitises the query ran against the database.