MGS-4853 Redirect to add trailing slash for page URLs
The canonical URL for pages includes a trailing slash,
and the intention is to redirect from pages accessed
without a trailing slash to add a trailing slash.
This is currently done by nginx, but it redirects
requests for resources such as publication downloads.

This change adds a SiteMapItemHandler to the sitemap
item for each type of page. It has intentionally not
been adding for URLs of the form: /isbn/* and
/Publications/* as these URLs will redirect anyway,
and there's no benefit in an intermediate redirect
to add a trailing slash.