MGS-4853 Use permanent redirect when adding trailing slash
The request URL is copied from a StringBuffer to a
StringBuilder, before appending the slash and, if
necessary, query string.
This is not for performance reasons, but because
the use of StringBuffer can be surprising (because
use of StringBuilder is more common now), and
SonarQube warns about its use. However, it's
unavoidable here, because the Servlet API uses it.