MGS-4846 Remove sitemap entry for site login page
Hippo provides a /login sitemap entry for the site by default. [1] This
is used for delivery tier authorisation - e.g. restricting a mount or
sitemap item to a specific role or group. [2]

We don't currently use this functionality. This commit removes the
ability to access this /login page and a see an 'authentication failed'
message.

[1] https://www.onehippo.org/library/concepts/hst-configuration-model/default-inherited-configuration.html
[2] https://www.onehippo.org/library/concepts/security/delivery-tier-authorization.html