Prevent exceptions logged by UrlAliasRedirectComponent
The data structure used to store redirects in the repo stores urls in a tree structure.  Nodes for which we have a redirect are stored in
the attribute govscot:url

Internal nodes may not have a url but previosuly we asumed that they did which caused an exception to be logged.  This adds a guard for
this case and also escapes the request path to ensure that it forms a valid jcr path.