MGS-4903 Override document text extraction class
For inclusion in the index, Hippo uses Apache Tika for text extraction,
and uses a special “hippo:text” property on the nodes with attachments
to prevent extraction occurring multiple times. We don’t wish to extract
text from our PDFs, since we have equivalent HTML content already and
don’t need to duplicate this.

Previously we could control this property using our migrations, but
since users have been uploading attachments directly using the CMS, we
no longer have that control.

To prevent Tika extraction, we need to override the class used. There's
no way to use a configuration file to change the behaviour, like we do
for the indexing configuration for example.

This is fragile and may require maintenance. It should be monitored
during upgrades.