MGS-5252 PublicationL:inkProcessor now returns 404 for bad links
PublicationsLinkProcessor had a link where it would pass on any url that began
with the path of a publication, e.g. /publication/mypublication/1

These kind of urls caused a NPE in the publication componenent and rather
than return a 404 the page was rendered as blank.  As a side efect this
also cause some redirects to break.