MGS-5264 Make display of primary document in ComplexDocument optional
Some complex documents do not have a primary document that should be given
prominence.  For cases like this we want a checkbox to control display
of the primary document.  New complex documents default to true.

A migration accompanies this commit which sets the value to true
for all docs besides those that comms have indicated.