MGS-5453 Add Statistics and research landing page
- Add the option for filtered list pages and side filter components to both be passed a
set of publication types in their component params (in the console). These publication types
must be valid ones existing in the publication types value list.
- Add bootstrapping for a Statistics and research filtered list page using these params.
- Add conditional to results component so list pages don't have to be an 'index' bean, as
the statistics and research list page content item will be in the same folder as the publications list page.
- Add bootstapping for new sitemap item for this page.
- Amend CSS for sidebar filter scrollable box to be 'max-height', so smaller lists only take up
as much height as is needed.