MGS-5453 disable the stats and research
the statistcs and research menu and landing page needs to be
be delayed as per digital comms request delayed. to do this,
we have removed the sitemap entry for the landing page and
removed the link from the site navigation. we have kept the
new site navigation styling from the pattern library