MGS-5642 Do not include contents page in redirect queries
The RedirectComponent queries for either Publicaiton on PubliocationPage items
that havei an old style gov.scot url recorded against it.  The first page of
the old publications sometimes had the same url as the publication (mostly
pre 2012).

This was causing multiple results to be returned and our code was logging a warning.
I have made the query more specific to resolve this.