MGS-5670 Fix handling of 404 for NewsComponent
The news component was not handling urls that did nto resolve to
a content item properly resulting in an NPE.

The 'fix' for this caused another problem due to the way it was serving
the 404 page. In fixing this I also realised that some legacy news
items dating to before PRGLoo became the supplier was not being handled
properly.