MGS-5875 improve event listener exception handling
For all of our event listeners / DaemonModule's we should ensure
that session.refresh(false) is called if a RepositoryException is
thrown in the course of handling the event.

To do this we now have a common base class called DaemonModuleBase
provides correct exception handling and also logs start and end
times.