Disable LookupFilter for site appender
This version of LookupFilter is incompatible with the version of
log4j in use (which contains security patches).
The result is that application logging does not work.

This change disables the filter, and can be removed once the
application has been upgraded to a version compatible with the
current version of log4j.