MGS-6085 Override user panel HTML
Since version 14's permissions overhaul, any user with user administration
permissions has the ability to change userroles for specific users. This allows
user admins to bypass any established group permissions, and allows for
privilege escalation.

To remedy this, we've added HTML in our CMS structure to override the default
markup, and to remove the userroles section of the user administration view.