MGS-6085 Override admin panel HTML
Since version 14's permissions overhaul, any user with user administration
permissions has the ability to change userroles for specific users. This allows
user admins to bypass any established group permissions, and allows for
privilege escalation.

This can also be done in the group admin view. We hide this panel from
appearing in the user administration menu, but it can still be reached by
clicking on groups that users are members of.

To remedy this, we've added HTML in our CMS structure to override the default
markup, and to remove the userroles section of the user and group
administration view.