MGS-7110 handle first page of a publication when pushing to funnelback
the first visible page of a publicaiton has the same canonical url as
the publication itself and so when pushing should use this as the key.
the logic has to take into account which page is the first visible page
i.e. the first page which is published and is not the contents pages