MGS-7481 Fix embargo permission escalation
Users in an embargo group were able to place any content under embargo,
regardless of their other permissions. This was because the embargo role
was defined in the wider content domain.

To fix this, the embargo roles have been moved to their specific
domains, so that users will only be able to embargo the content that
adhere to these domain rules.