MGS-7791 Fix list of minutes where two entries share the same meeting date
Minutes were being stored in a SortedSet using a date comparison to order
the set. This meant minutes with the same meeting date were considered
to be equal, and so the set could not store more than one set of minutes
with the same meeting date.

This change sorts minutes by meeting date, title and then canonical path
(in order to distinguish between any duplicates).