MGS-7780 prevent publicationType population causing performance issues
The previous implementation of PublicationTypeDerivedData caused the publication
type folder to be related to the new publication.  This had the unintended consquence
that when the type folder changed (for example when adding a new year folder), the
derived functions for all descending publivations were called causing the cms to lock up.

This  change makes the path of the new node available to the derived data function and
uses it to derive the type.  This prevents the relationship being created.

To do this we habe had to override an internal bloomreach class RelativePropertyReference.
We have raised a bloomreach ticket to have this added and should remove
this once it is available.
https://issues.onehippo.com/browse/CMS-16025