MGS-7694 SitemapEventListener to only change nodes specific to changed node
SitemapEventListener used to set the modified date of the year, month  and site
nodes in the data structure used to produce sitemap.xml's.  This could lead to
multiple concurrent changes throwing exceptions.  This would also cause changes
made by other event listeners to be rolled back.

This commit removes the maintenance of these dates and the caching logic in
SitemapResource.  To allow for this there is also now a scheduled daily task to
ensure that the